refactoring helper functions
