+++ lib/plugins/input/docker/dockerInspect.js
@@ -3,6 +3,7 @@ var Docker = require('dockerode')
 var docker = new Docker()
 var minimatch = require('minimatch')
 var consoleLogger = require('../../../util/logger.js')
+var parser = require('../../../util/parser.js')
 
 var dockerInfo = {}
 var tagIds = null
@@ -29,25 +30,6 @@ docker.info(function dockerInfoHandler (err, data) {
     }
   }
 })
-
-function parseReceiverList (receivers) {
-  let rv = []
-  if (!receivers) {
-    return undefined
-  }
-  let receiverList = receivers.split(',')
-  let parseUrlRegx = /(\S+:\/\/\S+?)\/(\S+)$/i
-  for (let i = 0; i < receiverList.length; i++) {
-    let url = receiverList[i].match(parseUrlRegx)
-    if (url && url.length === 3) {
-      rv.push({
-        url: url[1], // base url
-        index: url[2]
-      })
-    }
-  }
-  return rv
-}
 
 function getEnvVar (name, list) {
   if (!list) {
@@ -85,7 +67,6 @@ function getValue (name, list, info) {
           info.tags = {}
         }
         info.tags[keys[k]] = list[keys[k]]
-        info.tags
       }
     }
   } else {
@@ -173,7 +154,7 @@ function getLogseneEnabled (info) {
     logsReceiver = getEnvVar('LOGS_RECEIVER_URL', info.Config.Env)
   }
   if (logsReceiver) {
-    info.logsReceiver = parseReceiverList(logsReceiver)
+    info.logsReceiver = parser.parseReceiverList(logsReceiver)
   }
   // get optional logs target name
   if (info.Config && info.Config.Labels && info.Config.Labels.LOGS_DESTINATION) {
