+++ lib/plugins/output-filter/docker-log-enrichment.js
@@ -1,7 +1,8 @@
 var warningRegex = /warning/i
 var errorRegex = /[^|\S]error|exception/i
 var K8S = /^k8s_/
-var parseImageRegex = /^(\S+?\.\S+?\/|\S+?\:\d+\/){0,1}(\S+?):(\S+?){0,1}(@\S+?){0,1}$/
+
+var parser = require('../../util/parser.js')
 
 var k8sMetadata = {}
 if (process.env.SEVERITY_ERROR_PATTERN) {
@@ -46,24 +47,6 @@ function parseKubernetesInfo (containerName, logObject) {
   }
 }
 
-function parseImage (image) {
-  var rv = { name: image }
-  var result = parseImageRegex.exec(image)
-  if (result) {
-    if (result.length > 3) {
-      if (result[1]) {
-        rv.registry = result[1]
-      }
-      rv.name = result[2]
-      rv.tag = result[3]
-      if (result[4] !== undefined) {
-        rv.digest = result[4].substring(1, result[4].length)
-      }
-    }
-  }
-  return rv
-}
-
 module.exports = function enrichDockerLogs (context, config, eventEmitter, data, cb) {
   if (!context.container_name) {
     return cb(null, data)
@@ -72,7 +55,7 @@ module.exports = function enrichDockerLogs (context, config, eventEmitter, data,
     id: context.container_long_id,
     type: 'docker',
     name: context.container_name,
-    image: parseImage(context.image || ''),
+    image: parser.parseImage(context.image || ''),
     host: {
       hostname: process.env.SPM_REPORTED_HOSTNAME
     }
@@ -93,13 +76,6 @@ module.exports = function enrichDockerLogs (context, config, eventEmitter, data,
     context.elasticsearchUrl = context.dockerInspect.LOGS_RECEIVER_URL
   }
   if (context.labels) {
-    /*
-    var keys = Object.keys(context.labels)
-    for (var i=0; i < keys.length; i++) {
-      data[ keys[i] ] = context.labels[ keys[i] ]
-    }
-    */
-    console.log(context.labels)
     data.labels = context.labels
   }
